+++ backend/mqtt-handler/.env
@@ -1,0 +1,2 @@
+MQTT_BROKER_ADDR="tcp://localhost:1883"
+CONNECT_SERVER_ADDR="localhost:8553"
