+++ backend/mqtt-handler/README.md
@@ -2,4 +2,4 @@
 
 MQTT Brokerを経由してESP32に対して命令を出力、受け取るためのサービスです。
 
-やることとしては、基本的にProtobufで定義されたメッセージをMQTT BrokerにPublishすることです。
+やることとしては、基本的にProtobufで定義されたメッセージをMQTT BrokerにPublishするだけのStatelessなサービスです。
