+++ backend/mqtt-handler/pkg/handler/handler.go
@@ -1,0 +1,10 @@
+package handler
+
+func HandleMqttMessage(topic string, msg string) {
+	// TODO: Implement
+}
+
+func HandleConnectMessage(target string, msg string) {
+	// TODO: Implement
+	// mqttclient.Send(cc, "plarail2023/notify/train/arrival", "test")
+}
